Fixed bug in low-testing region
When testing low ratios of delta = m/N, we sometimes run into issues with the generation of degree sequences. If we set the number of tests per individual = 1, the sum over individuals may exceed the number of individuals in the groups when m/N is small. To fix this, we now randomly select among the largest indices in indeg and reduce them by one at each step until the sum(indeg) == sum(outdeg).